Importing current implementation of SLIC filter from SuperPixel module (#12)
* Importing current implementation of SLIC filter from SuperPixel module

* Remove override keyword at function definition

* Mark threadId as Unused to avoid warning

* Add common method to create cluster point

Create the joint domain cluster point from the image value and the
physical point location.

Note that vnl_vector_ref is derived from vnl_vector, enabling just the
implementation of vnl_vector for the argument.

* Adding class documentation

* Use Cluster and RefCluster typedefs

The vnl_vector_ref class is derived from the vnl_vector class. This
enables implementing methods for vnl_vector and passing vnl_vector_ref
objects as arguments. This is now conveyed with typedefs.

* Add VerifyInputInformation to verify output pixel type

Check that the output pixel type contains enough values to hold the
number of labels required.

* Only compute the residuals in debug mode

* Add time probe to profile filter execution time

* Adding proximity weight as a command line option

* Remove shadowing declaration of ImageDimension

* Adding comparison image for SLIC test

* Update cluster initialization algorithm

Change from usage of averaging bin shrink filter, to manual sampling
at "nearest neighbor" to evenly spacing with the user supergrid
spacing parameter.

* Style teaks

* Fix multi-threading of PerturbClusters method

There was a race condition with updating the clusters, and accessing
them. The threading is now performed over the indexes.

* Fix expecting the wrong number of threads

This produced errors in computation of the distribution of the work
load, and having some data not processed.

* Adding LabelPixelType for output image pixel

This should help make explicit casting clearer for the output label
image.

* fix unused variable warning

* Adding SLIC tests with 80 and 1 threads

There were multiple threading issues detected. The number 80 was
chosen because with that number of number of issue have been
detected.

* Add implementation to relabel based on connectivity

Implementation follow description in the paper with sequentially
relabeling connected components of sufficient size. For components of
insufficient size the prior labeled neighbor is used.

An optional multi-threaded enhancement has been added which keeps the
initial super-pixel component labels if the required criteria is met.

* Add robust initialization for algorithm variables

If there is an exception during processing, then algorithms variables
may not be cleaned up as expected. In the BeforeThreadedGenerateData
method, ensure that these variables are initialized to a know state
with clear contents.

* Adding printing of missing parameter variables

* Update Gradient perturb computation for scalars

The 1-norm of the Jacobian is used for the gradient magnitude to
perturb the cluster initial location.

* Add scalar support for Distance method

Using specialized dispatched methods for scalars and multi-component
images. The dispatching is done based on the assumption that for
scalars the ValueType is the same as the PixelType.

* Adding test for scalar SLIC

* Change cluster to be in index space

Avoid unneeded transform from index to physical space.

* Adding baseline images to scalar SLIC

* Add documentation to the filter's parameters

* Address unused variable warnings

* Improve function name and types used